Se finalizo el modulo rutas y usuarios

+++ frondend/src/app/mostrar-horario/mostrar-horario.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { BusService, Buses } from '../SERVICIOS/bus.service';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-mostrar-horario',
@@ -27,15 +27,29 @@ export class MostrarHorarioComponent implements OnInit {
   }
   eliminar(id:string)
   {
-    this.BusService.deleteBus(id).subscribe(
-      res=>{
-        console.log("ruta eliminada");
-        this.listarBus();
-      },
-      err=>console.log(err)
-    );
+    Swal.fire({
+      title: 'Esta seguro que desea eliminar el Bus?',
+      icon: 'warning',
+      showDenyButton: true,
+      confirmButtonText: `Eliminar`,
+
+      denyButtonText: `Cancelar`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.BusService.deleteBus(id).subscribe(
+          res=>{
+            console.log("bus eliminado");
+            this.listarBus();
+          },
+          err=>console.log(err)
+        );
+        Swal.fire('Se elimino el Bus', '', 'success')
+      }
+    })
+    
   }
   modificar(id:string){
+    
     this.router.navigate(['/Mostrar_Horario/Modificar_Horario/'+id])
   }
