+++ frondend/src/app/mostrar-rutas/mostrar-rutas.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { RutasService, Ruta } from '../SERVICIOS/rutas.service';
 import { Router, Routes } from '@angular/router';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-mostrar-rutas',
   templateUrl: './mostrar-rutas.component.html',
@@ -26,13 +27,26 @@ export class MostrarRutasComponent implements OnInit {
   }
   eliminar(id:string)
   {
-    this.RutasService.deleteRuta(id).subscribe(
-      res=>{
-        console.log("ruta eliminada");
-        this.listarRutas();
-      },
-      err=>console.log(err)
-    );
+    Swal.fire({
+      title: 'Esta seguro que desea eliminar la Ruta?',
+      icon: 'warning',
+      showDenyButton: true,
+      confirmButtonText: `Eliminar`,
+
+      denyButtonText: `Cancelar`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.RutasService.deleteRuta(id).subscribe(
+          res=>{
+            console.log("ruta eliminada");
+            this.listarRutas();
+          },
+          err=>console.log(err)
+        );
+        Swal.fire('Se elimino la Ruta', '', 'success')
+      }
+    })
+    
   }
   modificar(id:string){
     this.router.navigate(['/Mostrar_Rutas/Modificar_Ruta/'+id])
